Using which to detect files instead of test

+++ install.sh
@@ -13,10 +13,10 @@ cp ${REPO_CONFIG_DIR}/gitconfig ${HOME}/.gitconfig
 cp ${REPO_CONFIG_DIR}/hushlogin ${HOME}/.hushlogin
 
 # Make .dirs
-mkdir ${HOME_LOCAL_DIR} $2 >> /dev/null
-mkdir ${HOME_TB_DIR} $2 > /dev/null
-mkdir ${HOME_BIN_DIR} $2 > /dev/null
-mkdir ${HOME_BUILD_DIR} $2 > /dev/null
+mkdir ${HOME_LOCAL_DIR} 2> /dev/null
+mkdir ${HOME_TB_DIR} 2> /dev/null
+mkdir ${HOME_BIN_DIR} 2> /dev/null
+mkdir ${HOME_BUILD_DIR} 2> /dev/null
 
 # Copy tb scripts
 cp ${REPO_SCRIPT_DIR}/tb ${HOME_BIN_DIR}/tb
@@ -27,16 +27,44 @@ chmod 744 ${HOME_TB_DIR}/ip
 
 # Install dependencies
 
-echo "Installing Make"
-cd $HOME_BUILD_DIR
-git clone https://github.com/wkusnierczyk/make.git
-cd make
-./configure
-./make
-mv make $HOME_BIN_DIR
+if which -s make; then
+    echo "Make exists, skipping install"
+else
+    echo "Installing Make"
+    cd $HOME_BUILD_DIR
+    git clone https://github.com/wkusnierczyk/make.git
+    cd make
+    ./configure
+    ./make
+    mv make $HOME_BIN_DIR
+fi
 
-echo "Installing CMake"
+if which -s cmake; then
+    echo "CMake exists, skipping install"
+else
+    echo "Installing CMake"
+    cd $HOME_BUILD_DIR
+    git clone https://github.com/Kitware/CMake.git
+    cd CMake
+    ./bootstrap --prefix=${HOME_LOCAL_DIR} && make && make install
+fi
+
+
+if which -s neofetch; then
+    echo "Neofetch exists, skipping install"
+else
+    echo "Installing neofetch"
+    cd $HOME_BUILD_DIR
+    git clone https://github.com/dylanaraps/neofetch.git
+    cd neofetch
+    make PREFIX=${HOME_LOCAL_DIR} install
+fi
+
+# Clean build dir
+echo "Cleaning up the build directory"
 cd $HOME_BUILD_DIR
-git clone https://github.com/Kitware/CMake.git
-cd CMake
-./bootstrap --prefix=${HOME_LOCAL_DIR} && make && make install
+rm -rf ./*
+
+# Footer
+echo 
+echo "Install to ${HOME_LOCAL_DIR} is complete"
